pyproject.toml
[project]
name = "x-claw"
version = "0.1.0"
description = "X-Claw V2 skeleton (MiroThinker 30B first deployment)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "httpx>=0.27.0",
  "openai>=1.40.0",
  "PyYAML>=6.0.1",
  "prometheus-client>=0.20.0",
]

[project.optional-dependencies]
storage = [
  "redis>=5.0.0",
  "chromadb>=0.5.0",
]
mcp = [
  "mcp>=1.0.0",
]

[project.scripts]
x-claw = "x_claw.main:run"

[tool.uv]
dev-dependencies = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "ruff>=0.6.0",
]

